+++ src/bitbuffer.c
@@ -392,7 +392,7 @@ void bitrow_debug(uint8_t const *bitrow, unsigned bit_len)
     print_bitrow(bitrow, bit_len, 0, 1);
 }
 
-int bitrow_snprint(uint8_t const *bitrow, unsigned bit_len, char *str, unsigned size)
+int bitrow_snprint(uint8_t const *bitrow, unsigned bit_len, char *str, size_t size)
 {
     if (bit_len == 0 && size > 0) {
         str[0] = '\0';
