+++ src/confparse.c
@@ -73,7 +73,7 @@ char *readconf(char const *path)
         return NULL;
     }
 
-    off_t n_read = fread(conf, sizeof(char), file_size, fp);
+    size_t n_read = fread(conf, sizeof(char), file_size, fp);
     fclose(fp);
     if (n_read != file_size) {
         fprintf(stderr, "Failed to read \"%s\"\n", path);
